requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.128.4
filelock==3.20.3
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.9
langchain-core==1.2.9
langchain-huggingface==1.2.0
langchain-openai==1.1.7
langchain-openapi==0.1.1
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
MarkupSafe==3.0.3
openai==2.17.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
xxhash==3.6.0
zstandard==0.25.0
